MPITestImages

Documentation for MPITestImages.

MPITestImages.TestImage

MPITestImages.addOnTheFlyImage

MPITestImages.@testimage_gen

MPITestImages.TestImage — Type

Struct describing a testimage.

MPITestImages.addOnTheFlyImage — Method

Add a function symbol to the list of known test image generation functions.

MPITestImages.@testimage_gen — Macro

Macro for annotating functions that can be used to generate test images.