MPITestImages

Documentation for MPITestImages.

MPITestImages.TestImage

MPITestImages.addOnTheFlyImage

MPITestImages.changeScale

MPITestImages.checker_image

MPITestImages.delta_image

MPITestImages.testimage

MPITestImages.@testimage_gen

MPITestImages.TestImage — Type

Struct describing a testimage.

MPITestImages.addOnTheFlyImage — Method

Add a function symbol to the list of known test image generation functions.

MPITestImages.changeScale — Method

Loads the specified image from remote source and scales it accordingly.

MPITestImages.checker_image — Function

checker_image()
checker_image(size)
checker_image(size, checkersCount)
checker_image(size, checkersCount, stripeWidth)

Function to generate a phantom with a checker board pattern. This function uses a best effort approach, meaning that it is tried to cover most of the phantom with the pattern using the specified parameters.

Arguments

size::Tuple{Integer, Integer}: The size of the phantom

checkersCount::Tuple{Integer, Integer}: How many squares to generate along each axis

stripeWidth::Tuple{Integer, Integer}: By default (1, 1). Sets the width of the lines between the squares

Examples

julia> image = checker_image((8, 8), (2, 3), (2, 1))
8×8 Matrix{Float64}:
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  1.0  0.0  1.0  0.0  1.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  1.0  0.0  1.0  0.0  1.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0

MPITestImages.delta_image — Method

delta_image(size, numOfPoints; sizeOfPoint, distanceOfPoints, pivot, circularShape)

Function to generate a phantom with discrete points. The distanceOfPoints argument takes two functions that take the number of the point to generate and return an integer. This makes the phantoms to generate highly customizable.

Arguments

size::Tuple{Integer, Integer}: The size of the phantom

numOfPoints::Integer: The number of points to generate

sizeOfPoint::Tuple{Integer, Integer}: The size of the points in the phantom

distanceOfPoints::Tuple{Function, Function}: The distance to add between each points in x and y direction

pivot::Tuple{Integer, Integer}: The starting point to generate points towards (size, size)

circularShape::Bool: If true, points are generated as circular

Examples

Two simple dots

julia> image = delta_image((8, 8), 2; sizeOfPoint=(3, 2), distanceOfPoints=(x -> 0, x -> 4), pivot=(3, 3))
8×8 Matrix{Float64}:
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  1.0  1.0  0.0  0.0  1.0  1.0
 0.0  0.0  1.0  1.0  0.0  0.0  1.0  1.0
 0.0  0.0  1.0  1.0  0.0  0.0  1.0  1.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0

L-shaped arrangement

julia> image = delta_image((8, 8), 3; sizeOfPoint=(2, 2), distanceOfPoints=(x -> x == 2 ? 3 : 0, x -> x == 3 ? -3 : 3), pivot=(3, 3))
8×8 Matrix{Float64}:
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  1.0  1.0  0.0  0.0  0.0  0.0
 0.0  0.0  1.0  1.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  1.0  1.0  0.0  1.0  1.0  0.0
 0.0  0.0  1.0  1.0  0.0  1.0  1.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0

MPITestImages.testimage — Method

testimage(name, args; kwargs...)

Retrieve a test image with the given name and the matching parameters.

Note: The name must correspond either to a remote file or a function name annotated by the testimage_gen macro. If both exist, precedence is given to the function.

Examples

julia> image = testimage("delta_image", (8, 8), 2; sizeOfPoint=(3, 2), distanceOfPoints=(x -> 0, x -> 4), pivot=(3, 3))
8×8 Matrix{Float64}:
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  1.0  1.0  0.0  0.0  1.0  1.0
 0.0  0.0  1.0  1.0  0.0  0.0  1.0  1.0
 0.0  0.0  1.0  1.0  0.0  0.0  1.0  1.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0

MPITestImages.@testimage_gen — Macro

Macro for annotating functions that can be used to generate test images.